-- implement iterative approach to find transcript one by one -> gain speed

+++ src/tools/math_tools.h
@@ -5,7 +5,9 @@
 
 double gammaln(double x);
 
-double factln(int n);
+//double factln(int n);
+
+double factln(float n);
 
 double bino_pdf(int k, int n, float p);
